refactor: tx cpm[psables
